[GEOT-5740] Move DataUtilities URL methods to new URLs class in gt-metadata
