+++ docs/user/library/metadata/index.rst
@@ -46,3 +46,4 @@ The gt-metadata module is responsible for:
    geotools
    range
    text
+   urls
