+++ modules/library/main/src/main/java/org/geotools/data/DataUtilities.java
@@ -21,7 +21,6 @@ import java.io.File;
 import java.io.FileFilter;
 import java.io.FilenameFilter;
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Array;
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -29,7 +28,6 @@ import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.net.URLDecoder;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -62,13 +60,13 @@ import org.geotools.data.simple.SimpleFeatureLocking;
 import org.geotools.data.simple.SimpleFeatureReader;
 import org.geotools.data.simple.SimpleFeatureSource;
 import org.geotools.data.simple.SimpleFeatureStore;
+import org.geotools.data.view.DefaultView;
 import org.geotools.factory.CommonFactoryFinder;
 import org.geotools.factory.Hints;
 import org.geotools.feature.AttributeImpl;
 import org.geotools.feature.AttributeTypeBuilder;
 import org.geotools.feature.DefaultFeatureCollection;
 import org.geotools.feature.FeatureCollection;
-import org.geotools.feature.FeatureCollections;
 import org.geotools.feature.FeatureIterator;
 import org.geotools.feature.FeatureTypes;
 import org.geotools.feature.GeometryAttributeImpl;
@@ -88,11 +86,10 @@ import org.geotools.geometry.jts.ReferencedEnvelope;
 import org.geotools.geometry.jts.WKTReader2;
 import org.geotools.metadata.iso.citation.Citations;
 import org.geotools.referencing.CRS;
-import org.geotools.resources.i18n.ErrorKeys;
-import org.geotools.resources.i18n.Errors;
 import org.geotools.styling.UserLayer;
 import org.geotools.util.Converters;
 import org.geotools.util.NullProgressListener;
+import org.geotools.util.URLs;
 import org.geotools.util.Utilities;
 import org.opengis.coverage.grid.GridCoverage;
 import org.opengis.feature.Feature;
@@ -117,13 +114,11 @@ import org.opengis.filter.sort.SortBy;
 import org.opengis.filter.sort.SortOrder;
 import org.opengis.geometry.BoundingBox;
 import org.opengis.metadata.citation.Citation;
-import org.opengis.referencing.FactoryException;
 import org.opengis.referencing.ReferenceIdentifier;
 import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.opengis.util.ProgressListener;
 
 import com.vividsolutions.jts.geom.Coordinate;
-import com.vividsolutions.jts.geom.Envelope;
 import com.vividsolutions.jts.geom.Geometry;
 import com.vividsolutions.jts.geom.GeometryCollection;
 import com.vividsolutions.jts.geom.GeometryFactory;
@@ -134,8 +129,6 @@ import com.vividsolutions.jts.geom.MultiPoint;
 import com.vividsolutions.jts.geom.MultiPolygon;
 import com.vividsolutions.jts.geom.Point;
 import com.vividsolutions.jts.geom.Polygon;
-
-import org.geotools.data.view.DefaultView;
 
 /**
  * Utility functions for use with GeoTools with data classes.
@@ -181,7 +174,6 @@ import org.geotools.data.view.DefaultView;
  * </p>
  * And a grab bag of helpful utility methods for those implementing a DataStore:
  * <ul>
- * <li>{@link #urlToFile(URL)} and {@link #fileToURL(File)} and {@link #getParentUrl(URL)} used to work with files across platforms</li>
  * <li>{@link #includeFilters} and {@link #excludeFilters} work as a compound {@link FileFilter} making {@link File#listFiles} easier to use</li>
  * <li>{@link #propertyNames}, {@link #fidSet}, {@link #attributeNames} methods are used to double check a provided query and ensure
  * it can be correctly handed.</li>
@@ -201,8 +193,6 @@ public class DataUtilities {
     static Map<Class, String> typeEncode = new HashMap<Class, String>();
 
     static FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2();
-    
-    static final boolean IS_WINDOWS_OS;
 
     static {
         typeEncode.put(String.class, "String");
@@ -259,10 +249,6 @@ public class DataUtilities {
 
         typeEncode.put(Date.class, "Date");
         typeMap.put("Date", Date.class);
-        
-        // check if we are running on windows
-        String os = System.getProperty("os.name");
-        IS_WINDOWS_OS = os.toUpperCase().contains("WINDOWS");
     }
     
     /**
@@ -281,105 +267,17 @@ public class DataUtilities {
     }
 
     /**
-     * A replacement for File.toURI().toURL().
-     * <p>
-     * The handling of file.toURL() is broken; the handling of file.toURI().toURL() is known to be
-     * broken on a few platforms like mac. We have the urlToFile( URL ) method that is able to
-     * untangle both these problems and we use it in the geotools library.
-     * <p>
-     * However occasionally we need to pick up a file and hand it to a third party library like EMF;
-     * this method performs a couple of sanity checks which we can use to prepare a good URL
-     * reference to a file in these situtations.
-     * 
-     * @param file
-     * @return URL
+     * @deprecated Use {@link URLs#fileToUrl(File)}
      */
     public static URL fileToURL(File file) {
-        try {
-            URL url = file.toURI().toURL();
-            String string = url.toExternalForm();
-            if (string.contains("+")) {
-                // this represents an invalid URL created using either
-                // file.toURL(); or
-                // file.toURI().toURL() on a specific version of Java 5 on Mac
-                string = string.replace("+", "%2B");
-            }
-            if (string.contains(" ")) {
-                // this represents an invalid URL created using either
-                // file.toURL(); or
-                // file.toURI().toURL() on a specific version of Java 5 on Mac
-                string = string.replace(" ", "%20");
-            }
-            return new URL(string);
-        } catch (MalformedURLException e) {
-            return null;
-        }
+        return URLs.fileToUrl(file);
     }
 
     /**
-     * Takes a URL and converts it to a File. The attempts to deal with Windows UNC format specific
-     * problems, specifically files located on network shares and different drives.
-     * 
-     * If the URL.getAuthority() returns null or is empty, then only the url's path property is used
-     * to construct the file. Otherwise, the authority is prefixed before the path.
-     * 
-     * It is assumed that url.getProtocol returns "file".
-     * 
-     * Authority is the drive or network share the file is located on. Such as "C:", "E:",
-     * "\\fooServer"
-     * 
-     * @param url
-     *            a URL object that uses protocol "file"
-     * @return a File that corresponds to the URL's location
+     * @deprecated Use {@link URLs#urlToFile(URL)}
      */
     public static File urlToFile(URL url) {
-        if (!"file".equals(url.getProtocol())) {
-            return null; // not a File URL
-        }
-        String string = url.toExternalForm();
-        if( url.getQuery() != null){
-            string = string.substring(0, string.indexOf("?"));
-        }
-        if (string.contains("+")) {
-            // this represents an invalid URL created using either
-            // file.toURL(); or
-            // file.toURI().toURL() on a specific version of Java 5 on Mac
-            string = string.replace("+", "%2B");
-        }
-        try {
-            string = URLDecoder.decode(string, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException("Could not decode the URL to UTF-8 format", e);
-        }
-
-        String path3;
-
-        String simplePrefix = "file:/";
-        String standardPrefix = "file://";
-
-        if (IS_WINDOWS_OS && string.startsWith(standardPrefix)) {
-            // win32: host/share reference. Keep the host slashes.
-            path3 = string.substring(standardPrefix.length() - 2);
-            File f = new File(path3);
-            if (!f.exists()) {
-            	// Make path relative to be backwards compatible.
-            	path3 = path3.substring(2, path3.length());
-            }
-        } else if (string.startsWith(standardPrefix)) {
-            path3 = string.substring(standardPrefix.length());
-        } else if (string.startsWith(simplePrefix)) {
-            path3 = string.substring(simplePrefix.length() - 1);
-        } else {
-            String auth = url.getAuthority();
-            String path2 = url.getPath().replace("%20", " ");
-            if (auth != null && !auth.equals("")) {
-                path3 = "//" + auth + path2;
-            } else {
-                path3 = path2;
-            }
-        }
-
-        return new File(path3);
+        return URLs.urlToFile(url);
     }
 
     /**
@@ -2903,76 +2801,25 @@ public class DataUtilities {
     }
 
     /**
-     * Changes the ending (e.g. ".sld") of a {@link URL}
-     * 
-     * @param url
-     *            {@link URL} like <code>file:/sds/a.bmp</code> or
-     *            <code>http://www.some.org/foo/bar.shp</code>
-     * @param postfix
-     *            New file extension for the {@link URL} without <code>.</code>
-     * 
-     * @return A new {@link URL} with new extension.
-     * 
-     * @throws {@link MalformedURLException} if the new {@link URL} can not be created.
+     * @deprecated Use {@link URLs#changeUrlExt(URL, String)}
      */
-    public static URL changeUrlExt(final URL url, final String postfix)
+    public static URL changeUrlExt(URL url, String postfix)
             throws IllegalArgumentException {
-        String a = url.toExternalForm();
-        final int lastDotPos = a.lastIndexOf('.');
-        if (lastDotPos >= 0)
-            a = a.substring(0, lastDotPos);
-        a = a + "." + postfix;
-        try {
-            return new URL(a);
-        } catch (final MalformedURLException e) {
-            throw new IllegalArgumentException("can't create a new URL for " + url
-                    + " with new extension " + postfix, e);
-        }
+        return URLs.changeUrlExt(url, postfix);
     }
 
     /**
-     * The function is supposed to be equivalent to {@link File}.getParent(). The {@link URL} is
-     * converted to a String, truncated to the last / and then recreated as a new URL.
-     * 
-     * @throws {@link MalformedURLException} if the parent {@link URL} can not be created.
+     * @deprecated Use {@link URLs#getParentUrl(URL)}.
      */
-    public static URL getParentUrl(final URL url) throws MalformedURLException {
-        String a = url.toExternalForm();
-        final int lastDotPos = a.lastIndexOf('/');
-        if (lastDotPos >= 0)
-            a = a.substring(0, lastDotPos);
-
-        /**
-         * The parent of jar:file:some!/bar.file is jar:file:some!/, not jar:file:some!
-         */
-        if (a.endsWith("!"))
-            a += "/";
-
-        return new URL(a);
+    public static URL getParentUrl(URL url) throws MalformedURLException {
+        return URLs.getParentUrl(url);
     }
 
     /**
-     * Extends an {@link URL}.
-     * 
-     * @param base
-     *            Has to be a {@link URL} pointing to a directory. If it doesn't end with a
-     *            <code>/</code> it will be added automatically.
-     * @param extension
-     *            The part that will be added to the {@link URL}
-     * 
-     * @throws MalformedURLException
-     *             if the new {@link URL} can not be created.
+     * @deprecated Use {@link URLs#extendUrl(URL, String)
      */
     public static URL extendURL(URL base, String extension) throws MalformedURLException {
-        if (base == null)
-            throw new NullPointerException(Errors.format(ErrorKeys.NULL_ARGUMENT_$1, "base"));
-        if (extension == null)
-            throw new NullPointerException(Errors.format(ErrorKeys.NULL_ARGUMENT_$1, "extension"));
-        String a = base.toExternalForm();
-        if (!a.endsWith("/"))
-            a += "/";
-        a += extension;
-        return new URL(a);
+        return URLs.extendUrl(base, extension);
     }
 
     /**
