+++ modules/library/metadata/src/test/java/org/geotools/util/URLsTest.java
@@ -1,0 +1,108 @@
+/*
+ *    GeoTools - The Open Source Java GIS Toolkit
+ *    http://geotools.org
+ *
+ *    (C) 2017, Open Source Geospatial Foundation (OSGeo)
+ *
+ *    This library is free software; you can redistribute it and/or
+ *    modify it under the terms of the GNU Lesser General Public
+ *    License as published by the Free Software Foundation;
+ *    version 2.1 of the License.
+ *
+ *    This library is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *    Lesser General Public License for more details.
+ */
+
+package org.geotools.util;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+import java.io.File;
+import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URL;
+
+import org.junit.Test;
+
+/**
+ * Tests for {@link URLs}.
+ */
+public class URLsTest {
+
+    private void assertURL(String expectedFilePath, String urlString) throws MalformedURLException {
+        URL url = new URL(urlString);
+        File file = URLs.urlToFile(url);
+        if (URLs.IS_WINDOWS_OS) {
+            assertEquals(expectedFilePath.replaceAll("/", "\\\\"), file.getPath());
+        } else {
+            if (expectedFilePath.endsWith("/")) {
+                expectedFilePath = expectedFilePath.substring(0, expectedFilePath.length() - 1);
+            }
+            assertEquals(expectedFilePath, file.getPath());
+        }
+    }
+
+    private void handleFile(String path) throws Exception {
+        File file = new File(path);
+        URI uri = file.toURI();
+        URL url = file.toURI().toURL();
+        URL url2 = file.toURI().toURL();
+        assertEquals("jdk contract", file.getAbsoluteFile(), new File(uri));
+        File toFile = URLs.urlToFile(url);
+        assertEquals(path + ":url", file.getAbsoluteFile(), toFile);
+        File toFile2 = URLs.urlToFile(url2);
+        assertEquals(path + ":url2", file.getAbsoluteFile(), toFile2);
+    }
+
+    private String replaceSlashes(String string) {
+        return string.replaceAll("\\\\", "/");
+    }
+
+    @Test
+    public void testUrlToFile() throws Exception {
+        handleFile(System.getProperty("user.home"));
+        handleFile(System.getProperty("user.dir"));
+        if (URLs.IS_WINDOWS_OS) {
+            handleFile("C:\\");
+            handleFile("C:\\one");
+            handleFile("C:\\one\\two");
+            handleFile("C:\\one\\two\\and three");
+            handleFile("D:\\");
+            // Single slash keeps rooted path
+            assertURL("\\one", "file:/one");
+            assertURL("\\.\\one", "file:/./one");
+            // Double slash makes it relative if non existent.
+            assertURL("one", "file://one");
+            assertURL("./one", "file://./one");
+            handleFile("\\\\host\\share\\file");
+            // from GEOT-3300 urlToFile doesn't handle network paths correctly
+            URL url = new URL("file", "////oehhwsfs09", "/some/path/on/the/server/filename.nds");
+            File windowsShareFile = URLs.urlToFile(url);
+            assertNotNull(windowsShareFile);
+        } else {
+            handleFile("/one");
+            handleFile("one");
+            handleFile("/one/two/and three");
+            handleFile("/hello world/this++().file");
+        }
+        assertURL("one", "file:one");
+        assertURL("/one", "file:///one");
+        assertURL(replaceSlashes("C:\\"), "file://C:/");
+        assertURL(replaceSlashes("C:\\one"), "file://C:/one");
+        assertURL(replaceSlashes("C:\\one\\two"), "file://C:/one/two");
+        assertURL(replaceSlashes("C:\\one\\two\\and three"), "file://C:/one/two/and three");
+        assertEquals("sample", URLs.urlToFile(new URL("file:sample?query")).toString());
+        assertEquals("sample", URLs.urlToFile(new URL("file:sample#ref")).toString());
+        File file = File.createTempFile("hello", "world");
+        handleFile(file.getAbsolutePath());
+        handleFile(file.getPath());
+        // from GEOT-3300 urlToFile doesn't handle network paths correctly
+        URL url = new URL("file", "////oehhwsfs09", "/some/path/on/the/server/filename.nds");
+        File windowsShareFile = URLs.urlToFile(url);
+        assertNotNull(windowsShareFile);
+    }
+
+}
